observability_tracing_middleware/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
prometheus_client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-semantic-conventions==0.44b0
python-json-logger==2.0.7
